refactor: break out into separate components, import types
